feat: refactor imports to use types from '@/types/database' and '@/types/schema'
