+++ src/components/ChartVisualization.tsx
@@ -1,24 +1,7 @@
 import { useState } from "react";
-import {
-  BarChart,
-  Bar,
-  LineChart,
-  Line,
-  PieChart,
-  Pie,
-  ScatterChart,
-  Scatter,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  Legend,
-  ResponsiveContainer,
-  Cell,
-} from "recharts";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
-import { Download, BarChart3 } from "lucide-react";
+import { Download } from "lucide-react";
 import { toPng, toSvg } from "html-to-image";
 import { toast } from "sonner";
 import { ChartConfigPanel } from "./chart/ChartConfigPanel";
